requirements.txt
Pillow~=8.3.1
clipboard~=0.0.4
PyAudio~=0.2.11
Wave~=0.0.2
pynput~=1.7.3
PyQt5~=5.15.4